+++ src/component/Favourites/Favourites.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import PhotoItem from "../Body/PhotoItem";
+import PhotoItem from "../Gallery/PhotoItem";
 
 const Favourites = ({ photos, hoveredImageId, onMouseEnter, onMouseLeave }) => {
   const favouritesPhotos = photos.filter((photo) => photo.favourites);
@@ -7,13 +7,7 @@ const Favourites = ({ photos, hoveredImageId, onMouseEnter, onMouseLeave }) => {
   return (
     <div className="photos-container">
       {favouritesPhotos.map((photo) => (
-        <PhotoItem
-          key={photo.id}
-          photo={photo}
-          hoveredImageId={hoveredImageId}
-          onMouseEnter={onMouseEnter}
-          onMouseLeave={onMouseLeave}
-        />
+        <PhotoItem key={photo.id} photo={photo} />
       ))}
     </div>
   );
